+++ bonsai/src/bt.rs
@@ -28,7 +28,7 @@ pub struct BT<A, B> {
     /// Used by `RecordingTracer` to advance past unvisited subtrees in O(1).
     #[cfg(feature = "visualize")]
     #[cfg_attr(feature = "serde", serde(skip))]
-    pub(crate) node_metas: Vec<crate::telemetry::NodeMeta>,
+    pub(crate) node_metas: Vec<crate::tracer::NodeMeta>,
     /// Channel sender for shipping `TickTrace`s to the broadcaster thread.
     /// `None` when telemetry is not active or after the broadcaster has exited.
     #[cfg(feature = "visualize")]
@@ -53,7 +53,7 @@ impl<A: Clone, B> BT<A, B> {
         let bt = State::new(behavior);
 
         #[cfg(feature = "visualize")]
-        let node_metas = crate::telemetry::build_node_metas(&backup_behavior);
+        let node_metas = crate::tracer::build_node_metas(&backup_behavior);
 
         Self {
             state: bt,
@@ -106,11 +106,11 @@ impl<A: Clone, B> BT<A, B> {
             return self.tick_recording(e, f).map(|(result, _)| result);
         }
         self.tick_count += 1;
-        let mut tracer = crate::telemetry::NoopTracer;
+        let mut tracer = crate::tracer::NoopTracer;
         #[cfg(feature = "visualize")]
-        let metas: &[crate::telemetry::NodeMeta] = &self.node_metas;
+        let metas: &[crate::tracer::NodeMeta] = &self.node_metas;
         #[cfg(not(feature = "visualize"))]
-        let metas: &[crate::telemetry::NodeMeta] = &[];
+        let metas: &[crate::tracer::NodeMeta] = &[];
         match self.state.tick(0, metas, e, &mut self.bb, f, &mut tracer) {
             result @ (Status::Success | Status::Failure, _) => {
                 self.finished = true;
@@ -232,6 +232,10 @@ impl<A: Clone, B> BT<A, B> {
 
     /// Attach a live visualizer at `http://127.0.0.1:{port}/`.
     ///
+    /// Convenience for [`with_telemetry_at`](Self::with_telemetry_at) with the
+    /// loopback address. Use that method instead if you need to bind a
+    /// different host (e.g. `"0.0.0.0"` to expose the visualizer on the LAN).
+    ///
     /// Builder method: consumes and returns `Self` for chaining off [`BT::new`].
     /// Spawns a listener thread and a broadcaster thread. Telemetry is
     /// **best-effort** — [`TickTrace`](crate::telemetry::TickTrace)s are dropped
@@ -241,34 +245,8 @@ impl<A: Clone, B> BT<A, B> {
     /// After calling this method, [`tick`](Self::tick) automatically records and
     /// ships a trace on every call — no API change required.
     ///
-    /// Calling `with_telemetry` a second time replaces the existing sender; the
-    /// previous broadcaster exits on its next `recv` (sees `Disconnected`). If the
-    /// same port is still bound, the second bind returns `AddrInUse`.
-    ///
     /// # Errors
     /// Returns `io::Error` if `127.0.0.1:{port}` cannot be bound.
-    ///
-    /// # Edge cases
-    /// - **Cloning a telemetry-attached `BT` is not supported.** `BT` derives
-    ///   `Clone`, and `Option<SyncSender<_>>` clones share the underlying
-    ///   channel — both BTs would interleave their `TickTrace`s into the same
-    ///   broadcaster, producing nonsense in the visualizer. Either clone *before*
-    ///   calling `with_telemetry`, or call `with_telemetry` again on the clone
-    ///   (with a different port).
-    /// - **`reset_bt` preserves telemetry.** The sender survives, `tick_count`
-    ///   keeps climbing, `dropped_traces` resets to 0. Connected browsers see
-    ///   no disruption.
-    /// - **Already-finished BT.** `with_telemetry` still works — the tree
-    ///   definition is sent to clients on connect — but no `TickTrace`s flow
-    ///   until [`reset_bt`](Self::reset_bt) is called (`tick` returns `None`).
-    /// - **Deserializing a BT.** `telemetry_sender` is `#[serde(skip)]`, so the
-    ///   deserialized BT runs without telemetry until `with_telemetry` is called
-    ///   again — the typical "fresh run" workflow.
-    /// - **Pre-connect tick backlog.** Traces queue in the channel (capacity
-    ///   1024) until the first WS client connects. If no one connects within
-    ///   1024 ticks, older traces are dropped (`dropped_traces` increments);
-    ///   the first connecting client sees the tree definition plus the *next*
-    ///   tick onward, not the backlog.
     ///
     /// # Example
     /// ```no_run
@@ -279,7 +257,74 @@ impl<A: Clone, B> BT<A, B> {
     /// let mut bt = BT::new(behavior, bb).with_telemetry(8910)?;
     /// # Ok::<(), std::io::Error>(())
     /// ```
-    pub fn with_telemetry(mut self, port: u16) -> std::io::Result<Self>
+    pub fn with_telemetry(self, port: u16) -> std::io::Result<Self>
+    where
+        A: std::fmt::Debug,
+    {
+        self.with_telemetry_at("127.0.0.1", port)
+    }
+
+    /// Attach a live visualizer at `http://{addr}:{port}/`.
+    ///
+    /// Like [`with_telemetry`](Self::with_telemetry), but lets you pick the
+    /// bind address. Pass `"0.0.0.0"` to listen on every interface (so peers
+    /// on the LAN can connect), `"127.0.0.1"` for loopback only, or any
+    /// specific interface IP. The address is parsed by `TcpListener::bind`,
+    /// so hostnames and IPv6 literals (e.g. `"::1"`) also work.
+    ///
+    /// Spawns a listener thread and a broadcaster thread. Telemetry is
+    /// **best-effort** — [`TickTrace`](crate::telemetry::TickTrace)s are dropped
+    /// silently when the channel is full; the count is surfaced via
+    /// [`dropped_traces`](Self::dropped_traces).
+    ///
+    /// After calling this method, [`tick`](Self::tick) automatically records and
+    /// ships a trace on every call — no API change required.
+    ///
+    /// Calling either telemetry method a second time replaces the existing
+    /// sender; the previous broadcaster exits on its next `recv` (sees
+    /// `Disconnected`). If the same `(addr, port)` is still bound, the second
+    /// bind returns `AddrInUse`.
+    ///
+    /// # Errors
+    /// Returns `io::Error` if `{addr}:{port}` cannot be bound (invalid address,
+    /// `AddrInUse`, permission denied for low ports, etc.).
+    ///
+    /// # Edge cases
+    /// - **Cloning a telemetry-attached `BT` is not supported.** `BT` derives
+    ///   `Clone`, and `Option<SyncSender<_>>` clones share the underlying
+    ///   channel — both BTs would interleave their `TickTrace`s into the same
+    ///   broadcaster, producing nonsense in the visualizer. Either clone
+    ///   *before* attaching telemetry, or call `with_telemetry_at` again on
+    ///   the clone (with a different port).
+    /// - **`reset_bt` preserves telemetry.** The sender survives, `tick_count`
+    ///   keeps climbing, `dropped_traces` resets to 0. Connected browsers see
+    ///   no disruption.
+    /// - **Already-finished BT.** `with_telemetry_at` still works — the tree
+    ///   definition is sent to clients on connect — but no `TickTrace`s flow
+    ///   until [`reset_bt`](Self::reset_bt) is called (`tick` returns `None`).
+    /// - **Deserializing a BT.** `telemetry_sender` is `#[serde(skip)]`, so the
+    ///   deserialized BT runs without telemetry until a telemetry method is
+    ///   called again — the typical "fresh run" workflow.
+    /// - **Pre-connect tick backlog.** Traces queue in the channel (capacity
+    ///   1024) until the first WS client connects. If no one connects within
+    ///   1024 ticks, older traces are dropped (`dropped_traces` increments);
+    ///   the first connecting client sees the tree definition plus the *next*
+    ///   tick onward, not the backlog.
+    /// - **Binding `0.0.0.0`.** The URL printed to stdout is the address you
+    ///   passed; modern browsers map `http://0.0.0.0` to localhost, but for a
+    ///   peer to connect they need this machine's actual LAN IP.
+    ///
+    /// # Example
+    /// ```no_run
+    /// # use bonsai_bt::{Action, BT, Behavior::Sequence};
+    /// # use std::collections::HashMap;
+    /// let behavior = Sequence(vec![Action("step")]);
+    /// let bb: HashMap<String, i32> = HashMap::new();
+    /// // Expose to the LAN — anyone who can reach this host can view the tree.
+    /// let mut bt = BT::new(behavior, bb).with_telemetry_at("0.0.0.0", 8910)?;
+    /// # Ok::<(), std::io::Error>(())
+    /// ```
+    pub fn with_telemetry_at(mut self, addr: &str, port: u16) -> std::io::Result<Self>
     where
         A: std::fmt::Debug,
     {
@@ -287,13 +332,16 @@ impl<A: Clone, B> BT<A, B> {
         use std::net::TcpListener;
         use std::sync::mpsc::sync_channel;
 
-        let listener = TcpListener::bind(("127.0.0.1", port))?;
+        let listener = TcpListener::bind((addr, port))?;
+        // local_addr reflects the resolved socket — useful when port=0 (kernel-
+        // assigned) or when the user passes a hostname rather than a literal IP.
+        let bound = listener.local_addr()?;
         let definition = serde_json::to_string(&TreeDefinition::build(&self.initial_behavior))
             .expect("TreeDefinition is always serializable");
         let (tx, rx) = sync_channel::<TickTrace>(1024);
         crate::visualizer_server::spawn_server(listener, definition, rx)?;
         self.telemetry_sender = Some(tx);
-        println!("bonsai-bt visualizer: open http://127.0.0.1:{port}/");
+        println!("bonsai-bt visualizer: open http://{bound}/");
         Ok(self)
     }
 }
